# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FinTrack Multi-Company Finance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List all companies
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    patch:
      operationId: updateCompany
      tags: [companies]
      summary: Update a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"

  # Categories
  /companies/{companyId}/categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories for a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /companies/{companyId}/categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Transactions
...
  /companies/{companyId}/invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices for a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, sent, paid, overdue, cancelled]
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create an invoice
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /companies/{companyId}/invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get an invoice
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update an invoice
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete an invoice
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Receipts
...
components:
  schemas:
...
    Company:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        taxId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    CompanyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        taxId:
          type: string
    CompanyUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        taxId:
          type: string
    Category:
      type: object
      required: [id, companyId, name, type, createdAt]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [expense, deposit, both]
        color:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    CategoryInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
          enum: [expense, deposit, both]
        color:
          type: string
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
          enum: [expense, deposit, both]
        color:
          type: string
...
    InvoiceLineItem:
      type: object
      required: [description, quantity, unitPrice]
      properties:
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
    Invoice:
      type: object
      required: [id, companyId, invoiceNumber, clientName, issueDate, dueDate, status, lineItems, subtotal, taxRate, taxAmount, total, createdAt]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        invoiceNumber:
          type: string
        clientName:
          type: string
        clientEmail:
          type: ["string", "null"]
        clientAddress:
          type: ["string", "null"]
        issueDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
        subtotal:
          type: number
        taxRate:
          type: number
        taxAmount:
          type: number
        total:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    InvoiceInput:
      type: object
      required: [clientName, issueDate, dueDate, lineItems, taxRate]
      properties:
        invoiceNumber:
          type: string
        clientName:
          type: string
          minLength: 1
        clientEmail:
          type: string
        clientAddress:
          type: string
        issueDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
        taxRate:
          type: number
          minimum: 0
        notes:
          type: string
    InvoiceUpdate:
      type: object
      properties:
        invoiceNumber:
          type: string
        clientName:
          type: string
          minLength: 1
        clientEmail:
          type: string
        clientAddress:
          type: string
        issueDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
        taxRate:
          type: number
          minimum: 0
        notes:
          type: string
...